pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docfiler"
version = "0.1.0"
description = "AI-powered document filing CLI"
requires-python = ">=3.10"
dependencies = [
    "pikepdf>=10.0",
    "pdfminer.six>=20221105",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
docfiler = "docfiler.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
